+++ tests/test_polars_utils_data.py
@@ -5,6 +5,7 @@ from pathlib import Path
 
 from utils.column_values.categorical_column_values import (
     AscwdsFilteringRule,
+    AscwdsFilteringRule,
     CareHome,
     JobRoleFilteringRule,
 )